pyproject.toml
[project]
name = "one-bailey"
version = "0.1.0"
description = "Automated Trading System"
authors = [
    {name = "JIHOJU",email = "jihoju96@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "dynaconf (>=3.2.11,<4.0.0)",
    "pyupbit (>=0.2.34,<0.3.0)",
    "yfinance (>=0.2.61,<0.3.0)",
    "ta-lib (>=0.6.3,<0.7.0)",
    "pymongo (>=4.13.0,<5.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "beanie (>=1.29.0,<2.0.0)",
    "influxdb-client (>=1.49.0,<2.0.0)",
    "redis (>=6.1.0,<7.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "streamlit (>=1.45.1,<2.0.0)",
    "plotly (>=6.1.1,<7.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "aiohttp (>=3.12.0,<4.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "websocket-client (>=1.8.0,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "matplotlib (>=3.10.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
pytest-asyncio = "^0.26.0"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
mypy = "^1.15.0"
pre-commit = "^4.2.0"
jupyter = "^1.1.1"
ipython = "^9.2.0"
rich = "^14.0.0"
faker = "^37.3.0"
httpx = "^0.28.1"
pytest-mock = "^3.14.0"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["services", "libs"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "dist"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests", "services/mvp-trading-service/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=services/mvp-trading-service/src --cov-report=html"
